Added level changing functionality, health is recalculated based on level

+++ PSP-Strategy/PSP-Strategy/GyvybiuSkaiciavimasPaprastas.cs
@@ -6,12 +6,12 @@ namespace PSP_Strategy
 {
     public class GyvybiuSkaiciavimasPaprastas : IGyvybiuSkaiciavimas
     {
-        public double papildomosGyvybes(string klase, int gyvybes, string kilme)
+        public double papildomosGyvybes(string klase, double gyvybes, string kilme)
         {
             return gyvybes;
         }
 
-        public double veikejoGyvybes(int gyvybes, double sarvai, int lygis)
+        public double veikejoGyvybes(double gyvybes, double sarvai, int lygis)
         {
             return gyvybes;
         }
