+++ PSP-Strategy/PSP-Strategy/GyvybiuSkaiciavimasTobulinamas.cs
@@ -7,7 +7,7 @@ namespace PSP_Strategy
     public class GyvybiuSkaiciavimasTobulinamas : IGyvybiuSkaiciavimas
     {
         private double gyvybesPerskaiciuota;
-        public double papildomosGyvybes(string klase, int gyvybes, string kilme)
+        public double papildomosGyvybes(string klase, double gyvybes, string kilme)
         {
             gyvybesPerskaiciuota = gyvybes;
             if (klase == "Kovotojas" && kilme == "Maneja")
@@ -17,7 +17,7 @@ namespace PSP_Strategy
             return gyvybesPerskaiciuota;
         }
 
-        public double veikejoGyvybes(int gyvybes, double sarvai, int lygis)
+        public double veikejoGyvybes(double gyvybes, double sarvai, int lygis)
         {
             gyvybesPerskaiciuota = gyvybes;
